AutoMapping: Match rules concurrently by default
On a CPU with 32 threads this resulted in about 4x overal speedup with
my current tests. For the AutoMapping autotests though, the small
overhead would in some cases lead to a 200x slowdown (0.2ms instead of
0.001ms).

This change introduces a new rule-map option called `MatchInOrder`,
which when set to true will revert back to the old behavior of applying
each rule directly after a match is found. This disables the concurrency
entirely.