AutoMapping: Accumulate touched layers in AutoMap While Drawing
During AutoMap While Drawing, entire rule maps are skipped when they do
not have the changed layer as an input layer. As implemented, this
optimization broke the processing of rules that were indirectly affected
by modifications.

Now a list of touched tile layers is kept, against which each subsequent
rule map is checked.

Closes #3313