Fixed use of custom property types in global object types file
By default the object types are stored in a global location. From there,
they are loaded before the project, and as such they could not refer to
any property types defined in a project. When they did, they would
silently fall back to string or int properties.

While it would be preferred to use a project-specific object types file
if they refer to custom property types, I've fixed this issue by always
reloading the object types file after a project is loaded.

Closes #3301