Fixed some World Tool actions no longer having names
Regression in f9a806a43c858c4d4a7471c2eb1ac3c8028d5e9c.

I had removed the languageChanged() call since it appeared to be
superfluous and it caused a warning about calling a virtual function in
the constructor. However, it wasn't entirely unnecessary due to the
implementation in the base class. Resolved while still avoiding to call
a virtual function in the constructor.