Fixed flipping horizontally to not rotate objects by 180 degrees
This change fixes a few things about flipping objects. Most importantly,
flipping horizontally is no longer implemented as a 180-degree rotation,
and works as expected for tile objects, changing their horizontal
flipping flag.

It also fixes the calculation of the center around which the selected
objects are flipped for isometric maps and to take the tile object
alignment option into account.

The change leaves one case broken, which is the flipping of rotated
shape objects on isometric maps. These objects behave so strangely, with
their rotation happening in screen space after their isometric
projection, that I guess the only real option is to flip them in screen
space. For now shape objects are still flipped in pixel space.

Closes #1704
Closes #2660