Split up object types file type selection
This mainly resolves a usability issue on macOS, where the file type
selector was not visible at all so the fact that there even were
different formats was therefor hidden.

On GNOME the situation was similar, with the file type selector being
visible, but the extensions hidden.

Used the opportunity to make the JSON format get suggested as the
default.