AutoMapping: Introduced AutoMappingContext
This context stores temporary data specific to applying the active
AutoMappers to a target map. Previously this data was stored inside each
AutoMapper and hence the rule maps had to be reloaded to be applied to a
different map. Now this reloading is no longer necessary.

At the same time, the context stores any new tilesets, layers and
objects that may need to be added to the target map. They are no longer
added directly by the AutoMapper, which avoids the need to remove unused
tilesets and empty layers in a cleanup pass at the end.

This change also makes progress towards using the AutoMapper as part of
the preview or to be able to merge undo steps (see issue #2166), though
for now that is still not possible since tile layers in the target map
are still changed in-place.

Finally I figured out how to make QtConcurrent::blockingMapped compile
on Qt 5, making AutoMapping with Qt 5 almost twice as fast (matching the
Qt 6 speed). The functor was simplified by re-using the lambda
internally.