AutoMapping: AutoMap While Drawing now merges undo commands
Made the ChangeProperties undo command cloneable, so it doesn't prevent
merging.

The merging of the AutoMapping undo command is still not supported when
any objects are added or removed as a result of AutoMapping. To support
this, the AddMapObjects and RemoveMapObjects (or their base class
AddRemoveMapObjects) need to implement ClonableUndoCommand.

Closes #2166