AutoMapping: All per-rule options can now also be set on the map
When set on the map, they become the defaults for all rules defined on
that map.

Also, rule options now have to be set on either the map or the object to
become effective. Inherited values are no longer considered, because
that makes it easier to work with a custom "RuleOptions" object type,
which otherwise would cause all options set on the map to get overridden
even if you'd change only one of the properties.

Autotest has been adjusted to also set a rule property on the map, which
is then overridden for one of the rules.

RuleOptions struct now uses flags instead of std::optional, which is not
only more memory-efficient but also allows re-using the struct in places
where we don't need each member to be optional.