AutoMapping: Added "Forbid" match type
When a tile with "Forbid" match type is present, the "input" and
"inputnot" layers are effectively reversed. In practice, using this tile
will most of the time avoid the need for managing any "inputnot" layers.

It also makes the "NonEmpty" match type redundant, since it is the same
as "Empty" + "Forbid", but I think it makes sense to keep it as a
shortcut.