Improved snapping behavior for scalable objects on staggered maps
Now both the start position and the end position are snapped to the grid,
instead of the start position and the size. Especially on staggered maps,
snapping the size leads to confusing and undesired results.

This leads to a behavioral change for orthogonal and plain isometric maps, when
snapping is enabled after starting the creation of a scalable object. This no
longer makes the size a multiple of the snap distance. I doubt anybody would be
relying on this particular behavior, though.