Fixed map selection rectangle in world for infinite maps
In the world file, there is still only stored x,y,width,height for each
map, which is not really convenient for infinite maps, which extend into
negative coordinate space. Ideally their bounding rectangle would be
stored independently of their position.

For now, this change at least fixes the hover highlight and the moving
of infinite maps to not glitch around, but the hidden
onlyShowAdjacentMaps option will not work correctly.

Closes #3340