Always select first entry while using the Open File in Project action
Previously, it would keep the same entry selected when it was still
present after changing the filter text. This behavior was not useful
however, since when changing the filter text, the currently selected
entry is likely not the one you want and the more likely one is the
first one.