Simplified change object transform commands (#3349)
The previous implementation relied on changing objects directly and then 
creating an undo command when finishing up the move, resize or rotate
action. Now the undo command merges and becomes obsolete as expected, so
it can be used during the action.

This also affects changes to object position, size and rotation done 
through the Properties view.