tests: Fixed AutoMapping test for infinite map
To avoid breaking infinite maps again, like in c1d0ebd2c177b8374895345,
we need a functional test for infinite target maps. Whereas a the data
was added as part of the fix, the AutoMapping test still needed
adjustments to work with infinite maps.

This also exposed a bug in the TileLayer::iterator::key() and
TileLayer::const_iterator::key() functions, which hadn't been noticed
since so far they had apparently not been used.