Fixed restoring of window layout when maximized
In case of maximized windows, restoring the geometry causes two resize
events. We need to wait until the second resize event before trying to
restore the internal layout, otherwise it may be broken due to not
fitting in the initially smaller window.

I consider this more of a workaround to an issue in Qt, since ideally
there would only be a single resize event. I've added a 200ms timer that
makes sure the layout is always restored, even when a second resize
event is not sent.

Closes #590