Scripting: Layer.tintColor is now #ffffff when not set
Since it does not appear to be possible to tell apart an "invalid"
QColor value from "black" when it is passed to the JS engine, it is more
helpful to return "white" for the tint color, which matches the behavior
of the unset tint color.

In principle the tint color could default to white also in C++, but this
would require special-case reset logic for this property, which doesn't
seem worth it right now.