Link to QtSvg
Tiled needs QtSvg to be present since some of the icons use the SVG
format. Directly depending on its ensures it will be available.

It technically doesn't need to link to this library, since it should get
loaded through the SVG icon engine plugin, but doing so ensures that
macdeployqt will reliably copy over the SVG icon engine and image format
plugins.