Keep references between objects when copy/pasting
When a group of objects is pasted and some of the pasted objects
contained references to other pasted objects, these references are now
kept intact.

Previously, such references would keep referring to the original
objects, or get broken or refer to random objects, in case of pasting to
another map.

Closes #3361