Added usage options to custom classes
This adds options to limit where a custom class can be used. With this,
we can avoid clutter in the "Add Property" dialog, leaving out classes
which are not meant to be used as properties. And we can avoid clutting
the "Type" auto-suggest with classes that were not meant to be used as
object type.

While there is now the option to use a class property type as the type
for any kind of object, in practice this isn't possible yet until we
have moved the "type" property from MapObject/Tile up to Object.