Fixed locking up UI on property type name conflict
When an attempt to rename a property type by pressing Enter in the Name field
would cause a name conflict, two modal dialogs would pop up about the conflict,
and at least on macOS this would cause the UI to no longer accept any mouse
input.

This was due to editingFinished first getting emitted for the Enter key and
then also due to the loss of focus caused by popping up the first modal dialog.
I think this also highlights an issue in Qt, since editingFinished should not
get emitted a second time.

Closes #3380