Scripting: Fixed possible crash when accessing Layer.map
Actually seen on Sentry after exporting a map using a scripted export
format, since EditableMap::detachLayer also calls EditableLayer::map,
and in that case there was no asset or an invalid pointer.

I'm not entirely sure how that could happen, though, since there was an
EditableMap instance, which should be the asset of the EditableLayer.

Sentry-issue: TILED-4M