Qt 6.3: Avoid assertion when certain signals get disconnected
Since Qt 6.3, there is an assertion that triggers when a QObject derived
class gets destroyed and this in turn causes any signals to get emitted
that call back into the derived class after its destructor has executed.

This is indeed undesirable, so in a few cases I've added explicit
disconnects in the destructor, to avoid the connected slots from being
called after the destructor, which could otherwise lead to crashes.