Scripting: Work around missing enums in Qt object with Qt 6
In Qt 6, the enum values in the Qt object are only available in QML, or
in JS scripts used by QML. Plain JS files or modules are unfortunately
left without any way to import these values.

As a workaround, we can just copy the enum values from the QMetaObject
of the Qt namespace.

See https://codereview.qt-project.org/c/qt/qtdeclarative/+/418536/1 for
a possible solution on the Qt side.