Fixed shadow offset for other maps in a world
The painter scale is used to calculate the needed offset for shadows,
but since it is stored in the MapRenderer, we need to make sure this
value is updated for all maps displayed in a world.

This was a regression in b81264e66a3712827d4a6d, before which the scale
was being set by each tile layer item and object item.

Closes #3429