Fixed loading of custom tile image rectangles
The first time the image of a tile is set, the image rectangle was
getting initialized based on the image. However during loading, the
image rectangle is set before the image gets loaded.

Tweaked the check in Tile::setImage such that it will only initialize a
null image rect, or adjust the previous one if it exactly matched the
image, as before.

Closes #3405