Allow adding maps to image collection tilesets
Using maps as tiles in image collection tilesets was in general already
supported, but the UI ony allowed adding images.

Also adjusted the warning when an image or map fails to load, to have
just an OK button instead of Ignore/Cancel buttons, in case only one
file was selected.

Closes #3447