Hide actions from inactive tools from the search results
A few of the tools register their actions with the same Ids. They do
this because to the user there is only one such action, and they share
the any custom shortcut. When triggering these actions, we need to make
sure to trigger them for the right tool.

Implemented solution is to disable actions from inactive tools, and
search for the first enabled action for each registered Id.