Qbs: Use pkg-config to depend on zlib
This way we should support building Tiled with zlib in some strange
location, like on NixOS.

To support this on the latest version of Qbs, we need to set
Project.qbsModuleProviders since it no longer falls back on pkg-config
by default. Since this property was introduced in Qbs 1.21, this becomes
the minimum Qbs version.