Fixed possible crash on exit
It appears QNetworkAccessManager did not like to be destructed late,
which caused problems with the NewsFeed and NewVersionChecker classes
due to their usage of a static variable for the instance:

    QEventDispatcherWin32::wakeUp: Failed to post a message (Invalid window handle.)
    QSocketNotifier: Socket notifiers cannot be enabled or disabled from another thread

Now these classes are owned by the TiledApplication class instead, which
makes sure they get deleted earlier in the shutdown process.

Also using the QVersionNumber class now for comparing versions.