Fixed possible crash on activating search action
When using Qt 6 on Windows, the LocatorWidget could get destroyed as a
result of activating some actions, like when they open a file dialog.
This caused a crash because then after the action is triggered, close()
was called on a deleted object.

It should be no problem to call close() before triggering the action,
since the actual deletion of the LocatorWidget is delayed so
mLocatorSource should stay valid.