Reported a number of hard errors in the Console instead
There was little reason to fail the export in these cases, which are now
reported as errors or warnings in the Console view instead:

* WARNING: Unsupported layer type encountered
* WARNING: Only rectangles or polygons supported for collissions
* ERROR: Tile animations have gaps
* ERROR: Tile used that is also part of an animation
* ERROR: Unsupported tile flipping (120 degree or missing exportAlternates)

The choice between error or warning is a little arbitrary, since none of
these cause the export to fail and all of them cause the export to only
partially work. I picked warnings for cases where I think some people
may want to just ignore them.

The errors and warnings also enable jumping to the respective location,
layer or tile, where applicable.