Fixed painting issues caused by tile origin
The tile origin changes where the tiles are rendered, so it needs to be
taken into account when calculating the draw margins. These are used
when determining which area to repaint, and which tiles to render in a
certain area.

The draw margins are cached per map and now also per tileset. They are
invalidated when the "tile offset" or a tileset is changed as well as
when the "origin" of a tile is changed.