Fixed point object hover highlight position
The hover highlight for point objects did not render at the right
position for offset layers or layers with a custom parallax factor, when
the view was not scaled at 100%.

This was due to their transform not being updated to include that of the
view, which is disabled on purpose to keep the point markers at a
consistent size.

Fixed by including the layer offset in the position instead, since the
position is affected by the view scale as appropriate.

Closes #3571