Improved Shape Fill Tool handling of differently sized stamp variations
Rather than placing the variations on a fixed grid based on the maximum
stamp size, it now places variations next to each other horizontally and
also avoids creating entirely empty rows by determining the maximum
height of the variations placed on each row.

This can of course still cause gaps in the filled area when variations
of different height are placed next to each other, but avoiding that
entirely is more difficult or could be computationally too expensive.

Closes #3533