AutoMapping: Added support for output set probability
Rather than each output set always having equal probability, it is now
possible to specify a probability for each output set, using a custom
"Probability" property on any of its layers (when multiple layers in an
output set have a "Probability" property, the last / top-most one wins).

The probability between output sets is relative and default to 1.0. So
for example, when setting it to 2.0 for some output set, it will be
twice as likely to be picked relative to any others.

Closes #3179