Qbs: Added libDir option and use convenience install properties
Since Qbs 1.13, Application and DynamicLibrary items have convenient
"install" and "installDir" properties.

Added projects.Tiled.libDir option, which can be used to change the
directory libraries are installed to. This is useful for openSUSE
packaging (and possibly Fedora), which like to put 64-bit libraries in a
lib64 directory (see issue #3613).