JSON plugin: Fixed loading of empty tilesets created by script
When creating a new tileset using the scripting API, the tile size
defaults to 0x0. Previously, such a tileset would fail to load when
saved as JSON / TSJ since it didn't pass the validation of the tileset
parameters.

Since such a tileset has no image set, it is effectively an Image
Collection tileset, which actually does not need to have non-zero
"tilewidth" and "tileheight" properties. So the validation has been
slightly relaxed.

Closes #3542