Removed "Add Folder to Project" button from the startup page
This button is no longer very useful to show on this page and is now
even somewhat confusing. When creating a new project, its root folder is
added automatically after all.

The button also wasn't properly disabled when no project was loaded,
making it possible to add a folder before creating a project.