Scripting: Fixed behavior of Dialog.SameWidgetRows
When a new row was created automatically when adding a widget, it was
failing to remember the type of that widget, since addNewRow was called
after setting m_lastWidgetTypeName.

Now the last widget type is set in addDialogWidget instead. Also, we no
longer compare the type names but just compare the QMetaObject pointers
directly.

Also fixed a problem in ScriptDialog::deleteAllDialogs, which iterated
the sDialogInstances container while it was also getting modified as
each dialog was deleted.

Closes #3607