Refactor to share view panning code between MapView and TilesetView
Also fixes an issue where in the tileset view, the selection would keep
updating when triggering the panning with Space while selecting.

The "auto scrolling" mode is still not supported in the tileset view,
since this complicates things a little and I'm not sure if supporting
this in the tileset view is meaningful.