Fixed issue when corrections are not enabled
When corrections are disabled, and the Wang set is not complete, a
matching tile is only placed when it will not make it impossible to
place a matching tile on any of its neighboring cells.

For incomplete Wang sets with transitions to empty, it can happen that a
matching tile requires an empty WangId as neighbor. However, the empty
WangId is never explicitly part of a WangSet. As such, we should
probably never discard a match when it requires an empty neighbor.