Disabled placement of empty tiles by Stamp Brush in Terrain Fill mode
Also exposed the new TileLayerWangEdit.erasingEnabled property in the
scripting API.

Another change is that empty tiles no longer contribute to the desired
WangId for the Terrain Fill mode (when corrections are disabled).
Previously, only tiles outside of the map would not contribute to the
desired WangId. This change yields nicer behavior when using the Wang
Fill mode with the Stamp Brush, but it may need to be reconsidered for
the Shape Fill tool or a later Terrain Fill tool.