WangFiller: Don't try to make changes outside of a fixed map
This avoids the problem, that tiles set outside of the map by the
algorithm will affect the tiles inside the map, which is not desirable.

Also, explicitly set tiles that are outside of the map to empty.  This
seems like a good idea, since it means the preview will still highlight
those areas, but it will not affect the map once applied.