Fixed Terrain Fill Mode for sets containing transitions to empty (#3774)
Mostly removing the special handling for erasing and changing the way filling
is implemented. This does mean that terrain sets now support only 254
colors (down from 255).

* Slightly simplified updateAdjacent (one less call to "oppositeIndex").

* In wangIdFromSurrounding, prefer colors over "no color".

* In WangFiller::wangIdFromSurroundings, don't let empty tiles contribute to
  the WangId.

* Don't require tiles on the border of the region to match their surroundings,
  but only prefer them to.

* Slightly optimized WangBrush rotational symmetry mode.

* Disabled placement of empty tiles by Stamp Brush in Terrain Fill mode.

* Added TileLayerWangEdit.erasingEnabled property in the scripting API.

* WangFiller: Don't try to make changes outside of a fixed map.

This avoids the problem, that tiles set outside of the map by the algorithm
will affect the tiles inside the map, which is not desirable.

Also, explicitly set tiles that are outside of the map to empty. This seems
like a good idea, since it means the preview will still highlight those areas,
but it will not affect the map once applied.