Fixed crash when changing file member of custom class
This crash was caused by an infinite recursion happening when changing
top-level FilePath values. It only affected Qt 6 builds, because two
QVariant values that both hold a FilePath value always compared as
unequal due to FilePath not defining a comparision operator.

In addition to adding the comparison operator, I've also added a
condition to the updating of the property value. It should only be
necessary to update the top-level value when a nested value was changed.
Further more, it now sets mUpdatingDetails as another protection against
recursive calls.

For completeness I've also added the comparision operator to
PropertyValue and ObjectRef, though they would not have caused problems
in this context since they're not used as property display values.

Closes #3783