CI: Updated to Qbs 2.0.2
This should fix the macOS builds, which were using Qbs from Qt Creator
versions 5.0.3 and 7.0.2, which are no longer available to download.

Using Qt Creator 5.0.3 was necessary on old versions of macOS since it
was the last release using Qt 5. We should be able to use a newer
version even when building for macOS 10.12-10.13.