Scripting: Fixed Tool.updateEnabledState behavior
The setTargetLayerType(0) call in the ScriptedTool constructor happened
before the ScriptedTool instance was set as the prototype of the
provided script object. If the provided updateEnabledState would then
set the "this.enabled" property, it would override the Tool.enabled
property instead of changing it and cause all further attempts to change
the "this.enabled" property to not affect the Tool.enabled property.

Fixed by making sure the prototype is set before a possible call to
updateEnabledState happens.

Closes #3756