Fixed loading of invalid color properties
This was regression in d8b6b111ea6315c8fed9408227f65182c4bd52ef since it
used the return value of QVariant::convert to decide whether to use the
converted value or to keep the original (string) value. This, combined
with the fact that invalid colors are written out as empty strings,
broke the loading of invalid color values.

I seem to have added the check on the return value of convert to prevent
data loss when a conversion to the expected type fails. However, I'm not
entirely sure in which case this was preferable, so this change removes
the check again.

Also added an autotest, which can be adjusted or extended when we might
eventually run into the use-case for checking the "convert" return
value.

In the process, fixed the loadCircularReference test, which seems to
have been broken due to a change of behavior. And fixed the
loadAndSavePropertyTypes test, which broke due to adding "drawFill".

Closes #3793