JSON format: Fixed tile order when loading a tileset using the old format
Since Tiled 1.7, the order of tiles in the tileset file determines the
display order. But this should not be the case for tilesets stored in
the old JSON format (which was replaced in Tiled 1.2), since it used an
object with the tile IDs as string keys for storing tile meta-data and
these keys get sorted alphabetically.

This was resulting in tiles with associated data getting displayed in an
order like "1, 10, 11, 12, 13, ..., 2, 21, ...".