Draw map boundary so it doesn't overlap the map
Currently, the top and left border of the map overlap tile pixels, while the right and bottom do not. This change adjusts the border to always be outside the map.

This change offsets the top left corner by -1,-1 with the assumption that the border is always 1px wide because I could not find any code that ever sets the pen width. If the pen width is dynamic, then this offset will need to be adjusted for the width.