Use Q_GADGET to avoid needing to wrap things in QObject instances
Now QVector<WorldMapEntry> and QVector<WorldPattern> can be exposed to
JS directly. This only works with Qt 6 (tested with 5.15, 6.5 and 6.6),
but I don't think it's worth supporting these properties for Qt 5
builds.

Also removed World.displayName, since I think this was just a
convenience member which isn't very useful to add to the scripting API.
In the end it was just returning the file name part of the path returned
by World.fileName.

Improved the documentation in a few places.