Removed most world-specific signals
Removed worldLoaded, worldReloaded, worldUnloaded and worldSaved.

Made sure the following signals are also emitted for worlds:

* assetOpened (replaces worldLoaded)
* assetAboutToBeSaved
* assetSaved (replaces worldSaved)

Moved World.save() up to Asset.save(). Its behavior changed slightly, in
that errors are now reported in a modal dialog, due to reusing the
existing saving code.

There is no replacement for worldReloaded and worldUnloaded, but
hopefully worldsChanged suffices for most use-cases.

Fixed tiled.worldsChanged to only fire when the editor is loaded, not on
the CLI.