Scripting: Added Tile.image for accessing a tile's image data
Usually one could load the image based on Tile.imageFileName, but when a
tile's image has been set directly using Tile.setImage, this is not
possible. This can happen for example when an importer has used this
function since there was no external image file to refer to, or it used
an unsupported image format.

Tile.image now provides convenient access to the tile's image, falling
back to the Tileset's image when the tile has no individual image.

As such, this function also provides access to the tileset's image,
which would ideally be available through Tileset.image, but this
property was already taken and used for the image file name. To free up
this property in the future it is now deprecated. Extensions should
switch to the new Tileset.imageFileName property.

Part of issue #3630