Scripting: Restored compatibility for MapObject.polygon
While it seemed like a good idea to use the Qt 6 support for
transferring a QPolygonF to/from the JS engine, it caused a
compatibility breakage since the Qt 5 version was less strict on the
type.

The new version required the use of Qt.point to create the polygon's
points, whereas previously the points could also be given by
constructing a plain JS object like "{ x: 0, y: 0 }".

This change restores the behavior as of Tiled 1.10.1.

Closes #3845