Fixed updating of used tilesets after resizing map
When a tileset became unused due to a resize map or offset map
operation, the tileset could stay in the set of "used tilesets" reported
for that map.

This affected the scripting API Map.usedTilesets and caused an
unnecessary warning when trying to remove the affected tileset using the
"Remove Tileset" action.

Closes #3884