Automapping: Ignore empty outputs per-rule (#3896)
This change introduces a "compileOutputSet" step which is executed as
part of collecting the rules. In this step, each rule only receives as
possible outputs the output indexes that are non-empty for the region of
that rule.

This makes it much easier to add output variation to only some of the
rules. Previously, due to the possibility of empty outputs being chosen,
having rules with various amounts of variation required setting up
separate rule maps.

This new behavior not kick in for legacy rule maps that still define the
regions manually.

Also fixed a failing Automapping test (unrelated to this change).

Closes #3523