Fixed possible crash after assigning to tiled.activeAsset
When assigning a new map or tileset created from a script to tiled.activeAsset,
the necessary signals were not connected. This broke the signals from
EditableMap and it also caused certain ownership transfers from JS to C++ to
never take place. The latter could in turn lead to dangling pointers to objects
after they were deleted by the garbage collector.

Thanks to @eishiya for discovering this issue!