Terrain tools now ignore irrelevant labels
Changing the type of a Terrain Set does not clear no longer relevant
bits from each tile, in order to allow changing the type back and forth
without destroying this information.

However, this caused the terrain overlays to be confusing and terrain
tools to behave erratically. Now, the irrelevant bits of each tile's
terrain info is masked out where appropriate.

Closes #3204
Closes #3260