snap: Fixed determining of the version
When using the 'version: git' approach, only the final snap knows the
version, but it isn't known yet at compile time so Tiled always shows
its version as "git".

Setting fetch-depth to 0 makes it fetch all history for all branches and
tags, which is needed for `git describe` to work.