Display read-only state of open files using a lock icon
This should be mostly helpful for people using a VCS like Perforce or
Subversion, which may require a file to be checked out or locked before
it can be edited.

For now the read-only icon is a lock, which is common, but it may be
confusing for Subversion users since they actually need to lock the file
to make it writable.