Fixed repeated calls to findPluginSubclass
The problem was that handleError was not called when PyObject_IsSubclass
returned -1, which it commonly did when called with a value that wasn't
a class. With the error indicator still set, the next call to
PyObject_Dir was failing.

Since the errors aren't usually very useful, we just call PyErr_Clear when
PyObject_IsSubclass fails.

I couldn't find a convenient way to check if the value is a class, which
could avoid raising these errors in the first place.