Watch and reload all documents rather than just opened editors
When a map is loaded as part of a world, or a tileset is loaded as part
of a map, but not open in its own editor, previously its file was not
watched for changes.

Now these files are also watched. When a change is detected and no
unsaved changes are present, the affected map or tileset will be reloaded.

Closes #1785