Scripting: Fixed UI issues after setting class values
When setting a class value from a script, the value was not getting
converted from a JS object to a QVariantMap. This tripped up code
elsewhere, for example making the UI unable to edit the values (though
it could still display them, and they also got saved correctly).

This change adds the missing conversion from JS object to QVariantMap to
`tiled.propertyValue` and `Object.setProperty`.