Stamp Brush right-click cut when Shift is held
Implementing this required a refactor of the state system in
StampBrush.  Behavior and State were previously handled in the
same property, which resulted both in wrong behavior when using
capture and pressing the behavior keys (replicate by holding
right click, start holding shift, release right click, the behavior
wasn't set to line as it should have). This commit refactors this
system by splitting State and Behavior in two enums and two
properties, resulting in making the implemented feature possible,
fixing behavior bugs, and future-proofing this system for
expansion.

Implements #3961