Allow --project to be used alongside the --evaluate CLI parameter
It is useful to be able to load a project and its extensions when using
the --evaluate parameter to execute a particular script.

Also fixed a crash in Document constructor when using the --project
parameter, since it would generally cause a project to get loaded without
there being a DocumentManager.