Introduced "Docs" page
This makes it possible to more prominently link to the scripting API
docs, as well as to include a link to the rather helpful tip sheet by
eishiya.

It also avoids sending people unexpectedly to an entirely different
website when they just expect to click through some pages.

Also updated scripting API docs.