Implemented Remove and Reset buttons for custom properties
* Top-level custom properties can now be removed.
* Class members can be reset.

A "modified" state was added to Property, which is updated as necessary
and indicated by using bold font for the name.

Further tweaks to the spacing between property widgets and draw line to
separate property group headers.

A generic VariantMapProperty was split off from the CustomProperties
class, since the former can also be a base for editing class members.

The ColorButton widget now replaces its icon with "Unset" text when an
invalid color is set.

When changing the "Class" it now affects all selected objects.

Fixed layout flicker when collapsing nested class members.