Fixed issue with placing tile objects after switching maps
The "current tile" was not being restored correctly after switching
maps, when another type of object had been focused (for example due to
switching layers or placing objects).

Now, when there is no "current tile" referenced by the MapDocument, it
is set based on the tile selected in the current tileset view.

Closes #3681