Added a custom widget for the object flipping property
Using two toggle buttons rather than checkboxes, similar to the "Allowed
Transformations" tileset property.

Also changed the stretch factor of the property labels column to be the
same as the property editor column. It looks more balanced overall and
helps to keep the labels readable when propeties are nested a few
levels.