Remember the expanded state of custom class properties
The expanded state is now remembered for each "path", meaning you can
select different objects, and when they happen to have the same
properties the expanded state applies to all of them.

For now the state is lost when Tiled is closed. It might be nice to
store the list of expanded property paths in the session.