Group object "Position" and "Size" properties as "Geometry"
This is mostly to improve the layout since the W is often wider than the
X and Y.

The minimum size hint for the DoubleSpinBox was increased a bit so it
wraps a bit earlier.

This change also makes it so that changing any of these values changes
only that particular value for all selected objects.