CI: Updated builds to Qt 6.8.1 (#4110)
Replaced install-qt.sh with aqtinstall, since aqtinstall has already 
been patched to be able to install Qt 6.8.1. It appears to require
reading a destination path from an XML file, which is getting a bit much
for a shell script.

The Qt 6 AppImage builds are now done on Ubuntu 22.04, since the latest 
aqtinstall needed a more recent Python version. This required installing
libfuse2 explicitly, to allow the linuxdeploy AppImage to run.

The Qt 5 AppImage is now being built by Qbs 1.13. This is the oldest 
supported version of Qbs so it's good as additional test case.